// src/pages/Admin/Dashboard/Dashboard.type.ts
export type OverviewType = {
    currentAllTransition: number;
    beforeAllTransition: number;
    percentAllTransition: number;
    currentAllOrderPrice: number;
    beforeAllOrderPrice: number;
    percentAllOrderPrice: number;
    totalCustomer: number;
    currentAllNewGuest: number;
    beforeAllNewGuest: number;
    percentAllNewGuest: number;
};